fixing a potential overflow (#612)
[ROCm/rocSPARSEcommit: 99cf78b564fc984922b7029bfbf8299a6a9e3305]

+++ library/src/level2/rocsparse_bsrmv.cpp
@@ -514,7 +514,7 @@ rocsparse_status rocsparse_bsrmv_template(rocsparse_handle          handle,
         return rocsparse_status_success;
     }
 
-    ROCSPARSE_CHECKARG_ARRAY(8, nnzb * block_dim * block_dim, bsr_val);
+    ROCSPARSE_CHECKARG_ARRAY(8, int64_t(nnzb) * block_dim * block_dim, bsr_val);
     ROCSPARSE_CHECKARG_ARRAY(9, mb, bsr_row_ptr);
     ROCSPARSE_CHECKARG_ARRAY(10, nnzb, bsr_col_ind);
     ROCSPARSE_CHECKARG_POINTER(13, x);
